Database is working!

+++ src/main/scala/org/duckofdoom/howardbot/db/dto/User.scala
@@ -1,3 +1,3 @@
 package org.duckofdoom.howardbot.db.dto
 
-case class User(id: Int, userId: Int, username: String, firstName:String, lastName: String)
+case class User(id: Long, userId: Int, username: String, firstName:String, lastName: String)
